Added function IsFastForwarding to entitymanager
Intetnded to let systems check and see if this is a regular update,
or a fast-forwarded update.  They may have slightly different
behavior one way or another.

In particular, Sprite goes WAY FASTER if it doesn't have to
reconstruct the entire vertex buffer every update for a fast-forward
frame that will be never rendered...